Feat: 알람, 게시글 기능 추가 - [x] : 알림 생성 API 오류 수정 - [x] : 알림 조회 API 생성 - [x] : 참여 취소할 때 당일 취소는 불가 - [x] : 참여 취소 알림 생성 - [x] : 이미 참여 신청한 게시글 중복 불가

+++ src/main/java/com/gdsc/projectmiobackend/service/PostParticipationServiceImpl.java
@@ -2,17 +2,22 @@ package com.gdsc.projectmiobackend.service;
 
 import com.gdsc.projectmiobackend.common.ApprovalOrReject;
 import com.gdsc.projectmiobackend.dto.PostDto;
+import com.gdsc.projectmiobackend.entity.Alarm;
 import com.gdsc.projectmiobackend.entity.Participants;
 import com.gdsc.projectmiobackend.entity.Post;
 import com.gdsc.projectmiobackend.entity.UserEntity;
+import com.gdsc.projectmiobackend.repository.AlarmRepository;
 import com.gdsc.projectmiobackend.repository.ParticipantsRepository;
 import com.gdsc.projectmiobackend.repository.PostRepository;
 import com.gdsc.projectmiobackend.repository.UserRepository;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 
+import java.time.LocalDate;
+import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Objects;
 import java.util.stream.Collectors;
 
 @Service
@@ -21,6 +26,7 @@ public class PostParticipationServiceImpl implements PostParticipationService {
     private final PostRepository postRepository;
     private final UserRepository userRepository;
     private final ParticipantsRepository participantsRepository;
+    private final AlarmRepository alarmRepository;
 
     @Override
     public String participateInPost(Long postId, String email, String content) {
@@ -28,6 +34,10 @@ public class PostParticipationServiceImpl implements PostParticipationService {
                 .orElseThrow(() -> new IllegalArgumentException("유저정보가 없습니다."));
         Post post = postRepository.findById(postId)
                 .orElseThrow(() -> new IllegalArgumentException("Invalid Post ID: " + postId));
+
+        if(participantsRepository.findByPostIdAndUserId(postId, user.getId()) != null){
+            throw new IllegalArgumentException("이미 신청한 게시글입니다.");
+        }
 
         Participants participants = new Participants(post, user, content);
 
@@ -70,8 +80,18 @@ public class PostParticipationServiceImpl implements PostParticipationService {
         UserEntity user = userRepository.findByEmail(email)
                 .orElseThrow(() -> new IllegalArgumentException("유저정보가 없습니다."));
 
+        Post post = postRepository.findById(postId)
+                .orElseThrow(() -> new IllegalArgumentException("Invalid Post ID: " + postId));
+
+        if(Objects.equals(post.getTargetDate(), LocalDate.now())){
+            throw new IllegalArgumentException("당일 카풀은 취소할 수 없습니다.");
+        }
+
         Participants participants = participantsRepository.findByPostIdAndUserId(postId, user.getId());
 
+        Alarm alarm = new Alarm(LocalDateTime.now(), "신청을 취소한 사람이 있습니다!", post, participants.getUser());
+
+        alarmRepository.save(alarm);
         participantsRepository.delete(participants);
     }
 
@@ -106,7 +126,7 @@ public class PostParticipationServiceImpl implements PostParticipationService {
 
         participants.setApprovalOrReject(ApprovalOrReject.APPROVAL);
         post.setParticipantsCount(post.getParticipantsCount() + 1);
-        UserEntity user = participants.getUser();
+        //UserEntity user = participants.getUser();
 
 /*
         List<Participants> participants1 = participantsRepository.findByUserId(user.getId());
